+++ HISTORY.rst
@@ -1,6 +1,13 @@
 History
 =======
 
+1.1.2 (2020-02-15)
+------------------
+
+* Fixed issue of reconnection that device remains unavailable until state changes
+* Fixed retry code for strip type devices
+
+
 1.1.1 (2020-02-01)
 ------------------